Allow RP credentials to be used in plugin
The Authorizer comes in from the RP and is passed to the plugin for
subsequent requests. Before, the autorest.Authorizer object was created
using credentials harvested by the plugin.

In this PR, the fakerp package is also altered in order to pass along an
RP generated autorest.Authorizer mocked by enrichContext(). Otherwise,
the internal communication of the credentials is done via the
GetAuthorizerFromContext() routine.

Signed-off-by: Yannick Cote <ycote@redhat.com>